Fix permission on pid_dir and path to init scripts in Makefile

+++ Makefile
@@ -6,9 +6,9 @@ build:
 
 install: dist
 	pip3 -V
-	cp -n policyd_rate_limit/policyd-rate-limit.conf /etc/
-	cp -n policyd_rate_limit/init/policyd-rate-limit /etc/init.d
-	cp -n policyd_rate_limit/init/policyd-rate-limit.service /etc/systemd/system/ || true
+	[ ! -f /etc/policyd-rate-limit.conf ] && cp -n policyd_rate_limit/policyd-rate-limit.conf /etc/ || true
+	cp -n init/policyd-rate-limit /etc/init.d
+	cp -n init/policyd-rate-limit.service /etc/systemd/system/ || true
 	pip3 install policyd-rate-limit -U -f ./dist/policyd-rate-limit-${VERSION}.tar.gz
 	systemctl daemon-reload
 uninstall:
